Power iteration ok
wip for slepc

+++ opendiff/materials.cpp
@@ -6,17 +6,16 @@
 
 #include "materials.h"
 
-
 namespace mat
 {
-    Materials::Materials(const std::vector<Eigen::ArrayXXf> &values, const std::vector<std::string> &names,
+    Materials::Materials(const std::vector<Eigen::ArrayXXd> &values, const std::vector<std::string> &names,
                          const std::vector<std::string> &reac_names)
     {
         if (names.size() != values.size())
             throw std::invalid_argument("The number of values is != from the number of names!");
 
         // get the number of groups
-        getNbGroups(values); 
+        getNbGroups(values);
 
         // set the reactions properly
         setReactionsNames();
@@ -26,7 +25,7 @@ namespace mat
 
         // for each mat: create an empty eigen array with the size of m_reac_names times the number of energy groups
         int i = 0;
-        for (Eigen::ArrayXXf mat : values)
+        for (Eigen::ArrayXXd mat : values)
         {
             auto new_mat = addRemovalXS(mat, ids);
 
@@ -34,9 +33,7 @@ namespace mat
             m_values[names[i]] = new_mat;
             i++;
         }
-
     }
-
 
     std::vector<int> Materials::checkReacNamesOrder(const std::vector<std::string> &reac_names)
     {
@@ -48,7 +45,8 @@ namespace mat
         int i = 0;
         for (std::string reac_name : m_reac_names)
         {
-            if ( reac_name == "SIGR" ){
+            if (reac_name == "SIGR")
+            {
                 i++;
                 continue;
             }
@@ -56,12 +54,13 @@ namespace mat
             int ref_i = std::distance(reac_names.begin(), std::find(reac_names.begin(),
                                                                     reac_names.end(), reac_name));
 
-            if (static_cast<std::vector<int>::size_type>(ref_i) == reac_names.size()){
+            if (static_cast<std::vector<int>::size_type>(ref_i) == reac_names.size())
+            {
                 std::cerr << reac_name << "\n";
                 throw std::invalid_argument("There is a missing reaction!");
             }
             else
-                ids[ref_i] = i ;
+                ids[ref_i] = i;
 
             i++;
         }
@@ -69,11 +68,10 @@ namespace mat
         return ids;
     }
 
-
-    void Materials::getNbGroups(const std::vector<Eigen::ArrayXXf> &values)
+    void Materials::getNbGroups(const std::vector<Eigen::ArrayXXd> &values)
     {
         int i = 0;
-        for (Eigen::ArrayXXf mat : values)
+        for (Eigen::ArrayXXd mat : values)
         {
             // check array shape
             if (i == 0)
@@ -95,12 +93,12 @@ namespace mat
 
         m_reac_names.push_back("SIGR");
 
-        int i = 0 ;
-        for (auto reac_name : m_reac_names){
+        int i = 0;
+        for (auto reac_name : m_reac_names)
+        {
             m_reac2id[reac_name] = i;
-            i++ ;
+            i++;
         }
-
     }
 
     const double Materials::getValue(const std::string &mat_name, const int i_grp, const std::string &reac_name) const
@@ -108,14 +106,14 @@ namespace mat
         return m_values.at(mat_name)(i_grp, m_reac2id.at(reac_name));
     }
 
-    Eigen::ArrayXXf Materials::addRemovalXS(const Eigen::ArrayXXf &mat, const std::vector<int> &ids)
+    Eigen::ArrayXXd Materials::addRemovalXS(const Eigen::ArrayXXd &mat, const std::vector<int> &ids)
     {
         // check array shape
         if (static_cast<std::vector<int>::size_type>(mat.cols()) != ids.size())
             throw std::invalid_argument("The number of reaction is different in the array and the name's list!");
 
         // reorder the array
-        Eigen::ArrayXXf new_mat = mat(Eigen::all, ids);
+        Eigen::ArrayXXd new_mat = mat(Eigen::all, ids);
 
         // add the removal xs section
         new_mat.conservativeResize(mat.rows(), mat.cols() + 1);
@@ -138,23 +136,22 @@ namespace mat
                 new_mat(grp_orig, id_sigr) += new_mat(grp_orig, id_grp_dest);
             }
         }
-        return new_mat ; 
+        return new_mat;
     }
 
-    void Materials::addMaterial(const Eigen::ArrayXXf &mat, const std::string &name, const std::vector<std::string> &reac_names)
+    void Materials::addMaterial(const Eigen::ArrayXXd &mat, const std::string &name, const std::vector<std::string> &reac_names)
     {
         // check that the reac names are in the same order than m_reac_names
         auto ids = checkReacNamesOrder(reac_names);
 
-        //check the number of groups
+        // check the number of groups
         if (m_nb_groups != mat.rows())
             throw std::invalid_argument("The number of groups is not the same for all the materials!");
 
-        auto new_mat = addRemovalXS(mat, ids) ; 
+        auto new_mat = addRemovalXS(mat, ids);
 
         // fill the matrix
         m_values[name] = new_mat;
     }
 
 } // namespace mat
-
