+++ opendiff/materials.h
@@ -14,7 +14,7 @@ namespace mat
     {
     private:
         // vector of eigen matrix
-        std::map<std::string, Eigen::ArrayXXf> m_values{};
+        std::map<std::string, Eigen::ArrayXXd> m_values{};
 
         // we add the transfert section in it (always at the end) and  "SIGR"
         std::vector<std::string> m_reac_names{"D", "SIGA", "NU_SIGF", "CHI"};
@@ -25,17 +25,16 @@ namespace mat
 
         std::vector<int> checkReacNamesOrder(const std::vector<std::string> &reac_names);
         void setReactionsNames();
-        void getNbGroups (const std::vector<Eigen::ArrayXXf> &values);
-        Eigen::ArrayXXf addRemovalXS(const Eigen::ArrayXXf &mat, const std::vector<int> &ids);
-
+        void getNbGroups(const std::vector<Eigen::ArrayXXd> &values);
+        Eigen::ArrayXXd addRemovalXS(const Eigen::ArrayXXd &mat, const std::vector<int> &ids);
 
     public:
-        Materials() = delete ;
+        Materials() = delete;
         Materials(const Materials &copy) = default;
-        Materials(const std::vector<Eigen::ArrayXXf> &values, const std::vector<std::string> &names,
+        Materials(const std::vector<Eigen::ArrayXXd> &values, const std::vector<std::string> &names,
                   const std::vector<std::string> &reac_names);
 
-        const Eigen::ArrayXXf & getMaterial(const std::string &name) const { return m_values.at(name); };
+        const Eigen::ArrayXXd &getMaterial(const std::string &name) const { return m_values.at(name); };
         const auto getMaterials() const { return m_values; };
 
         const double getValue(const std::string &mat_name, const int i_grp, const std::string &reac_name) const;
@@ -44,12 +43,9 @@ namespace mat
         const auto getMatNames() const { return m_values; };
         const int getNbGroups() const { return m_nb_groups; };
         const int getReactionIndex(const std::string &reac_name) const { return m_reac2id.at(reac_name); };
-        void addMaterial(const Eigen::ArrayXXf &mat, const std::string &name, const std::vector<std::string> &reac_names);
-
+        void addMaterial(const Eigen::ArrayXXd &mat, const std::string &name, const std::vector<std::string> &reac_names);
     };
 
 } // namespace mat
-
-
 
 #endif // MATERIALS_H
